fixed fundWallet function

+++ contracts/interfaces/curve/ICurveDeposit.sol
@@ -36,7 +36,7 @@ interface ICurveDeposit {
     /// @dev returns swap contract from new curve deposit zap
     function pool() external view returns (address);
 
-    function base_coins(int128 arg0) external view returns (address);
+    function base_coins(uint256 arg0) external view returns (address);
 
     function underlying_coins(int128 arg0) external view returns (address);
 }
